Move project root stuff to project-root module
