+++ src/state.ts
@@ -1,11 +1,10 @@
 import * as vscode from 'vscode';
 import Analytics from './analytics';
 import * as util from './utilities';
-import * as config from './config';
 import * as path from 'path';
 import * as os from 'os';
-import * as fs from 'fs';
 import { getStateValue, setStateValue } from '../out/cljs-lib/cljs-lib';
+import * as projectRoot from './project-root';
 
 let extensionContext: vscode.ExtensionContext;
 export function setExtensionContext(context: vscode.ExtensionContext) {
@@ -144,55 +143,12 @@ export async function initProjectDir(uri?: vscode.Uri): Promise<void> {
   } else {
     const workspaceFolder = getProjectWsFolder();
     console.log(workspaceFolder);
-    const candidatePaths = await findProjectRootPaths();
-    const closestRootPath = findClosestProjectRootPath(candidatePaths);
-    const projectRootPath = await pickProjectRootPath(candidatePaths, closestRootPath);
+    const candidatePaths = await projectRoot.findProjectRootPaths();
+    const closestRootPath = await projectRoot.findClosestProjectRootPath(candidatePaths);
+    const projectRootPath = await projectRoot.pickProjectRootPath(candidatePaths, closestRootPath);
     setStateValue(PROJECT_DIR_KEY, projectRootPath);
     setStateValue(PROJECT_DIR_URI_KEY, vscode.Uri.file(projectRootPath));
   }
-}
-
-async function findProjectRootPaths() {
-  const projectFileNames: string[] = ['project.clj', 'shadow-cljs.edn', 'deps.edn'];
-  const projectFilesGlob = `**/{${projectFileNames.join(',')}}`;
-  const excludeDirsGlob = `**/{${config.getConfig().projectRootsSearchExclude.join(',')}}`;
-  const t0 = new Date().getTime();
-  const candidateUris = await vscode.workspace.findFiles(projectFilesGlob, excludeDirsGlob, 10000);
-  console.debug('glob took', new Date().getTime() - t0, 'ms');
-  const projectFilePaths = candidateUris.map((uri) => path.dirname(uri.fsPath));
-  const candidatePaths = [...new Set(projectFilePaths)].sort();
-  console.log({ candidatePaths });
-  return candidatePaths;
-}
-
-function findClosestProjectRootPath(candidatePaths: string[]) {
-  const doc = util.tryToGetDocument({});
-  console.log(doc);
-  const docDir = doc && doc.uri ? path.dirname(doc.uri.fsPath) : undefined;
-  const closestRootPath = docDir
-    ? candidatePaths
-        .filter((p) => docDir.startsWith(p))
-        .sort()
-        .reverse()[0]
-    : candidatePaths[0];
-  console.log(closestRootPath);
-  return closestRootPath;
-}
-
-async function pickProjectRootPath(candidatePaths: string[], closestRootPath: string) {
-  const pickedRootPath =
-    candidatePaths.length < 2
-      ? undefined
-      : await util.quickPickSingle({
-          title: 'Project root',
-          values: candidatePaths,
-          default: closestRootPath,
-          placeHolder: 'Multiple Clojure projects found, pick one',
-          saveAs: `projectRoot`,
-          autoSelect: true,
-        });
-  const projectRootPath = candidatePaths.includes(pickedRootPath) ? pickedRootPath : closestRootPath;
-  return projectRootPath;
 }
 
 /**
